Replace hashed passwords with session secrets
- Updated ARCHITECTURE.md for password terminology changes.
- Replaced `sessionPassword` with `sessionSecret` across the app.
- Merged hashing and HMAC utilities into passwordCrypto.
- Removed password hashing in favor of direct secret use.
- Simplified password handling in StreamContext, hooks, and utils.
