--- app/src/utils/passwordHasher.ts
@@ -1,49 +1,0 @@
-const encoder = new TextEncoder();
-
-const bufferToHex = (buffer: ArrayBuffer): string => {
-  const bytes = new Uint8Array(buffer);
-  return Array.from(bytes).map(b => b.toString(16).padStart(2, '0')).join('');
-};
-
-const hexToBuffer = (hex: string): ArrayBuffer => {
-  const normalized = hex.trim().toLowerCase();
-  if (normalized.length % 2 !== 0) {
-    throw new Error('Invalid hex string length');
-  }
-
-  const buffer = new ArrayBuffer(normalized.length / 2);
-  const bytes = new Uint8Array(buffer);
-  for (let i = 0; i < normalized.length; i += 2) {
-    const byte = Number.parseInt(normalized.slice(i, i + 2), 16);
-    if (Number.isNaN(byte)) {
-      throw new Error('Invalid hex string value');
-    }
-    bytes[i / 2] = byte;
-  }
-  return buffer;
-};
-
-export async function hashPassword(password: string): Promise<string> {
-  const data = encoder.encode(password);
-  const hashBuffer = await crypto.subtle.digest('SHA-256', data);
-  return bufferToHex(hashBuffer);
-}
-
-export async function hmacSha256(keyHex: string, message: string): Promise<string> {
-  const keyData = hexToBuffer(keyHex);
-  const key = await crypto.subtle.importKey(
-    'raw',
-    keyData,
-    { name: 'HMAC', hash: 'SHA-256' },
-    false,
-    ['sign']
-  );
-  const signature = await crypto.subtle.sign('HMAC', key, encoder.encode(message));
-  return bufferToHex(signature);
-}
-
-export function generateNonce(byteLength: number = 16): string {
-  const bytes = new Uint8Array(byteLength);
-  crypto.getRandomValues(bytes);
-  return Array.from(bytes).map(b => b.toString(16).padStart(2, '0')).join('');
-}
